Split out BP seeder; bit more testing
